requirements.txt
anyio==3.5.0
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
backports.cached-property==1.0.1
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
fastapi==0.75.1
flake8==4.0.1
graphql-core==3.2.1
h11==0.13.0
httptools==0.4.0
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysk-utils @ git+https://github.com/suankularb-wittayalai-school/mysk-utils.git@main
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
Pygments==2.11.2
pyparsing==3.0.8
pytest==7.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
strawberry-graphql==0.106.2
tomli==2.0.1
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.2
zipp==3.8.0
